refactor: App에서 useShoppingCart 훅으로 상태 분리
